cext/stubs: fix 4 clippy regressions from msgpack spike (PR #46)
Master CI red on four lints that landed via the cext-msgpack
stubs in PR #46. Same shapes I fixed in PR #45's L3-D json batch
— the msgpack PR added new functions with copies of the old
anti-patterns. Fixes match the PR #45 idioms:

- dispatch.rs:224  `b"\\0".as_ptr() as *const c_char` → `c"".as_ptr()`
  (C-string literal)
- dispatch.rs:263  `unsafe { crate::Qtrue }` (Qtrue is a const,
  no unsafe op inside) → drop the unsafe block
- strings.rs:291-293  `N as *mut c_void` (encoding sentinel
  fallback in rb_enc_from_index) → `without_provenance_mut(N)`
- types.rs:163  `n as i64` on c_longlong (rb_ll2inum) → drop
  the cast; c_longlong is guaranteed ≥64 bits per C standard

+++ crates/rubyrs-cext/src/stubs/types.rs
@@ -158,9 +158,11 @@ pub unsafe extern "C" fn RHASH_SIZE(v: Value) -> c_long {
 // ===== msgpack-ruby additions =====
 
 /// long long -> Integer. rubyrs's Number is i64; identical to rb_ll2num.
+/// `c_longlong` is guaranteed ≥64 bits per the C standard so `n` is
+/// already i64 on every target — no widening cast needed.
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn rb_ll2inum(n: c_longlong) -> Value {
-    with_state(|st| st.intern(CValue::Int(n as i64)))
+    with_state(|st| st.intern(CValue::Int(n)))
 }
 
 /// unsigned long long -> Integer. Truncates to i64 (rubyrs has no
